requirements.txt
ddddocr==1.4.7
Flask==2.2.2
numpy==1.23.2
Pillow==9.2.0
Requests==2.31.0
selenium==4.10.0
